Major cleanup and changes to structure for clarity.
