--- Food+Love/Controllers/Main/AdmirersVC.swift
@@ -1,18 +1,0 @@
-//  AdmirersVC.swift
-//  Food+Love
-//  Created by Gloria Washington on 3/21/18.
-//  Copyright © 2018 Gloria Washington. All rights reserved.
-
-
-import UIKit
-
-class AdmirersVC: UIViewController {
-
-	override func viewDidLoad() {
-		super.viewDidLoad()
-
-	}
-
-
-
-}
