+++ Food+Love/Controllers/Main/DatesVC.swift
@@ -1,30 +1,17 @@
 //
-//  PlanDateVC.swift
+//  DateVC.swift
 //  Food+Love
 //
-//  Created by C4Q on 3/13/18.
-//  Copyright © 2018 Winston Maragh. All rights reserved.
+//
 //
 
 import UIKit
 
 class DatesVC: UIViewController {
-
     override func viewDidLoad() {
-        super.viewDidLoad()
-
+			super.viewDidLoad()
     }
 
-    
 
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }
