+++ Food+Love/Controllers/Main/ProfileVC.swift
@@ -1,7 +1,11 @@
+//
 //  ProfileVC.swift
 //  Food+Love
-//  Created by C4Q on 3/13/18.
-//  Copyright © 2018 Winston Maragh. All rights reserved.
+//
+//
+//
+
+
 
 import UIKit
 import Firebase
