+++ Food+Love/Models/Message.swift
@@ -1,3 +1,4 @@
+
 //  Message.swift
 //  Food+Love
 //  Created by Winston Maragh on 3/16/18.
