+++ Food+Love/Models/Restaurant Search/FSSearch.swift
@@ -1,16 +1,15 @@
-//
-//  FourSquareSearchJSON.swift
+
+//	FSSearchJSON.swift
 //  Food+Love
-//
-//  Created by C4Q on 3/14/18.
+//  Created by Winston Maragh on 3/14/18.
 //  Copyright © 2018 Winston Maragh. All rights reserved.
-//
 
 import Foundation
 
 
+//////////////////////////////////////////
 // MARK: FourSquare Search Json
-struct FourSquareSearchJSON: Codable {
+struct FSSearchJSON: Codable {
 	let response: SearchResponse
 }
 
@@ -69,6 +68,36 @@ struct Delivery: Codable {
 	let provider: Provider?
 	struct Provider: Codable {
 		let name: String?
+	}
+}
+
+
+//////////////////////////////////////////
+// MARK: - FourSquare Search API Client
+struct FSSearchAPIClient {
+	private init(){}
+	static let manager = FSSearchAPIClient()
+
+	func getVenues(from search: String, coordinate: String?, near: String?, completion: @escaping (Error?, [Venue]?) -> Void) {
+
+		var endpoint = ""
+		if let near = near, near != "" {endpoint = "https://api.foursquare.com/v2/venues/search?near=\(near)&query=\(search)\(FourSquareAPIKeys.fourSquareAuthorization)"}
+		else if let coordinate = coordinate {endpoint = "https://api.foursquare.com/v2/venues/search?ll=\(coordinate)&query=\(search)\(FourSquareAPIKeys.fourSquareAuthorization)"}
+
+		guard let url = URL(string: endpoint) else {return}
+
+		let task =  URLSession.shared.dataTask(with: url, completionHandler: { (data, response, error) in
+			if let error = error {completion(error, nil)}
+			else if let data = data {
+				do {
+					let JSON = try JSONDecoder().decode(FSSearchJSON.self, from: data)
+					let venues = JSON.response.venues
+					completion(nil, venues)
+				}
+				catch { print("Error processing data: \(error)") }
+			}
+		})
+		task.resume()
 	}
 }
