+++ Food+Love/Models/Restaurant Search/YelpAPIKeys.swift
@@ -1,9 +1,11 @@
+
 //  YelpAPIKeys.swift
 //  Food+Love
-//  Created by C4Q on 3/14/18.
+//  Created by Winston Maragh on 3/14/18.
 //  Copyright © 2018 Winston Maragh. All rights reserved.
 
 import Foundation
+
 
 struct YelpAPIKeys {
 	static let todaysDate = Date().description.prefix(10).replacingOccurrences(of: "-", with: "")
