+++ Food+Love/Services/ImageService.swift
@@ -1,12 +1,12 @@
+
 //  ImageService.swift
 //  Food+Love
-//  Created by C4Q on 3/14/18.
+//  Created by Winston Maragh on 3/14/18.
 //  Copyright © 2018 Winston Maragh. All rights reserved.
 
-//
-//import Foundation
-//
-//
+import Foundation
+
+
 //Image Helper - get images from online
 //struct ImageService {
 //	private init() {}
