+++ Food+Love/Services/NetworkHelper.swift
@@ -1,10 +1,12 @@
+
 //  NetworkHelper.swift
 //  Food+Love
-//  Created by C4Q on 3/14/18.
+//  Created by Winston Maragh on 3/14/18.
 //  Copyright © 2018 Winston Maragh. All rights reserved.
 
 import Foundation
 import UIKit
+
 
 //HTTP
 enum HTTPVerb: String {
