+++ Food+Love/SupportingFiles/AppDelegate.swift
@@ -1,3 +1,4 @@
+
 //  AppDelegate.swift
 //  Food+Love
 //  Created by C4Q on 3/13/18.
@@ -18,19 +19,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 		FirebaseApp.configure()
 
-		//	//Navigation Bar
-			UINavigationBar.appearance().backgroundColor = UIColor.blue
+		//Navigation Bar
 			UINavigationBar.appearance().tintColor = UIColor.white
 			UINavigationBar.appearance().alpha = 1.0
 			UINavigationBar.appearance().titleTextAttributes = [.foregroundColor : UIColor.white]
 
-		//
-		//	//Tab Bar
-		//	UITabBar.appearance().backgroundColor = UIColor.white
+		//Tab Bar
 			UITabBar.appearance().tintColor = UIColor.white
 			UITabBar.appearance().alpha = 1.0
 			UITabBarItem.appearance().setTitleTextAttributes([.foregroundColor: UIColor.white], for: .normal)
-		//
 
 		return true
 	}
