+++ Food+Love/Views/Welcome Views/WelcomeLogoSlide.swift
@@ -1,9 +1,11 @@
+
 //  WelcomeLogoSlide.swift
 //  Food+Love
 //  Created by Winston Maragh on 3/20/18.
 //  Copyright © 2018 Winston Maragh. All rights reserved.
 
 import UIKit
+
 
 class WelcomeLogoSlide: UIView {
