+++ Food+Love/Views/Welcome Views/WelcomeSlide.swift
@@ -1,9 +1,11 @@
-//  Slide.swift
+
+//  WelcomeSlide.swift
 //  Food+Love
 //  Created by Winston Maragh on 3/17/18.
 //  Copyright © 2018 Winston Maragh. All rights reserved.
 
 import UIKit
+
 
 class WelcomeSlide: UIView {
